### 2024-02-26 #229 @ewallace
@labarba thank you, this is great news! I've read through the paper and don't have any immediate updates. The carpentries lab reviews covered the materials and lesson, but did not review the paper.md. So I'll look forward to editorial comments from @JasonJWilliamsNY and be ready to revise in response to those.
