### 2024-02-27 b287d11
site deploy
Auto-generated via {sandpaper}
Source  : 53c342c71e971556adc7e767edeb96533fd50bd1
Branch  : md-outputs
Author  : GitHub Actions <actions@github.com>
Time    : 2023-09-23 22:02:33 +0000
Message : markdown source builds

Auto-generated via {sandpaper}
Source  : 7db847ebaf4460e94599bdc3a3ff9c224c14967c
Branch  : main
Author  : Allen Lee <alee@users.noreply.github.com>
Time    : 2023-09-23 21:58:42 +0000
Message : Merge pull request #138 from comses-education/fix_naming_sorting_challenge_formatting

fix challenge formatting
