+++ pkgdown.yml
@@ -2,5 +2,5 @@ pandoc: 2.19.2
 pkgdown: 2.0.7
 pkgdown_sha: ~
 articles: {}
-last_built: 2024-02-20T01:10Z
+last_built: 2024-02-27T01:11Z
 
