+++ pkgdown.yml
@@ -1,6 +1,6 @@
-pandoc: 2.19.2
+pandoc: 3.1.11
 pkgdown: 2.0.7
 pkgdown_sha: ~
 articles: {}
-last_built: 2024-02-27T01:11Z
+last_built: 2024-03-05T01:11Z
 
