+++ md5sum.txt
@@ -5,7 +5,7 @@
 "index.md" "d1e47c29006d4fd14db51a9785916fde" "site/built/index.md" "2023-08-31"
 "paper.md" "437862826f96a03774d19249189ac16c" "site/built/paper.md" "2023-09-13"
 "episodes/01-introduction.md" "20db3572b403389eccd97c3425717354" "site/built/01-introduction.md" "2023-08-31"
-"episodes/02-data_management.md" "ec17a220df1259ef60ed84f8fbffd3b2" "site/built/02-data_management.md" "2023-09-07"
+"episodes/02-data_management.md" "7f3fd13aafe24b8b7d6b46cfa1be76a2" "site/built/02-data_management.md" "2024-03-05"
 "episodes/03-software.md" "7dc30bc0c113045c390ba3c8e4fcbf94" "site/built/03-software.md" "2023-08-31"
 "episodes/04-collaboration.md" "2bd0f509102efb43ad311055a381e475" "site/built/04-collaboration.md" "2023-09-23"
 "episodes/05-project_organization.md" "c2a58b5bbc3e45493582c1527b3f82e2" "site/built/05-project_organization.md" "2023-09-23"
