+++ md5sum.txt
@@ -1,7 +1,7 @@
 "file" "checksum" "built" "date"
 "CODE_OF_CONDUCT.md" "c93c83c630db2fe2462240bf72552548" "site/built/CODE_OF_CONDUCT.md" "2023-08-31"
 "LICENSE.md" "b24ebbb41b14ca25cf6b8216dda83e5f" "site/built/LICENSE.md" "2023-08-31"
-"config.yaml" "5aabf494762d8a815eb74a273db2ccbf" "site/built/config.yaml" "2023-09-07"
+"config.yaml" "192dcf2d94675dfb593bc0fc2e8b2426" "site/built/config.yaml" "2024-03-07"
 "index.md" "d1e47c29006d4fd14db51a9785916fde" "site/built/index.md" "2023-08-31"
 "paper.md" "437862826f96a03774d19249189ac16c" "site/built/paper.md" "2023-09-13"
 "episodes/01-introduction.md" "20db3572b403389eccd97c3425717354" "site/built/01-introduction.md" "2023-08-31"
