### 2024-03-07 #229 @labarba
@kyleniemeyer — I've invited you to steer this JOSE submission through the fast-track publication, as it has already been through the Carpentries review. Can you help us out? (@JasonJWilliamsNY : might you kindly give Kyle a little summary of the fast-track process?)
