+++ pkgdown.yml
@@ -2,5 +2,5 @@ pandoc: 3.1.11
 pkgdown: 2.0.7
 pkgdown_sha: ~
 articles: {}
-last_built: 2024-03-12T01:11Z
+last_built: 2024-03-19T01:12Z
 
