### 2024-03-19 #229 @kyleniemeyer
Hi @ewallace, the work here is just focused on the paper. I think the presence of the `date` field is causing the compilation to fail (just my guess, though). Could you remove that?
