### 2024-04-02 32e8227
site deploy
Auto-generated via {sandpaper}
Source  : 1874bc5ce6b5f8904c03bf4c35a28bcb52b05784
Branch  : md-outputs
Author  : GitHub Actions <actions@github.com>
Time    : 2024-03-05 07:56:05 +0000
Message : markdown source builds

Auto-generated via {sandpaper}
Source  : db169b15bf1cde5a40dbfe0bbb43a746fcd2a873
Branch  : main
Author  : Allen Lee <alee@users.noreply.github.com>
Time    : 2024-03-05 07:55:14 +0000
Message : Merge pull request #145 from comses-education/update_relative_links

fix: replace absolute links with relative links
