+++ pkgdown.yml
@@ -2,5 +2,5 @@ pandoc: 3.1.11
 pkgdown: 2.0.7
 pkgdown_sha: ~
 articles: {}
-last_built: 2024-03-26T01:11Z
+last_built: 2024-04-02T01:16Z
 
