+++ pkgdown.yml
@@ -2,5 +2,5 @@ pandoc: 3.1.11
 pkgdown: 2.0.7
 pkgdown_sha: ~
 articles: {}
-last_built: 2024-04-02T01:16Z
+last_built: 2024-04-09T01:13Z
 
