+++ pkgdown.yml
@@ -1,6 +1,6 @@
 pandoc: 3.1.11
-pkgdown: 2.0.7
+pkgdown: 2.0.8
 pkgdown_sha: ~
 articles: {}
-last_built: 2024-04-09T01:13Z
+last_built: 2024-04-16T01:13Z
 
