+++ pkgdown.yml
@@ -1,6 +1,6 @@
 pandoc: 3.1.11
-pkgdown: 2.0.8
+pkgdown: 2.0.9
 pkgdown_sha: ~
 articles: {}
-last_built: 2024-04-16T01:13Z
+last_built: 2024-04-23T01:15Z
 
