### 2024-05-07 de265f5
site deploy
Auto-generated via {sandpaper}
Source  : 05f1b83e5962e4e8f774581d4f70f5b745d46384
Branch  : md-outputs
Author  : GitHub Actions <actions@github.com>
Time    : 2024-04-24 20:09:25 +0000
Message : markdown source builds

Auto-generated via {sandpaper}
Source  : 4a478c3606accc1deead313d63ad7342359d25c7
Branch  : main
Author  : Allen Lee <alee@users.noreply.github.com>
Time    : 2024-04-24 20:08:16 +0000
Message : Merge pull request #146 from carpentries-lab/126-zenodo-doi

add zenodo DOI to config.yaml
