+++ pkgdown.yml
@@ -1,6 +1,5 @@
 pandoc: 3.1.11
-pkgdown: 2.0.9
+pkgdown: 2.1.0
 pkgdown_sha: ~
 articles: {}
-last_built: 2024-07-02T01:22Z
-
+last_built: 2024-07-09T01:22Z
