+++ pkgdown.yml
@@ -2,4 +2,4 @@ pandoc: 3.1.11
 pkgdown: 2.1.0
 pkgdown_sha: ~
 articles: {}
-last_built: 2024-07-09T01:22Z
+last_built: 2024-07-16T01:24Z
