+++ pkgdown.yml
@@ -1,5 +1,5 @@
 pandoc: 3.1.11
-pkgdown: 2.1.0
+pkgdown: 2.1.1
 pkgdown_sha: ~
 articles: {}
-last_built: 2024-09-17T01:16Z
+last_built: 2024-09-24T01:42Z
