### 2024-10-11 carpentries-lab/good-enough-practices#150 @emma-wilson
Paper review
Hi @ewallace, I noticed over on the JOSE GitHub (https://github.com/openjournals/jose-reviews/issues/229) there's been a hold up on the review because the PDF of the paper isn't compiling. Just wanted to check this was on your radar? (because this totally dropped off my radar for months!)
