### 2025-04-23 #229 @labarba
I'd feel a bit bad withdrawing the paper without giving @ewallace a chance to pop in here and say wassup. We all drop plates and forget we had them spinning. I have so many that I have dropped and neglected myself!
