### 2025-04-23 carpentries-lab/good-enough-practices#150 @ewallace
Thanks @emma-wilson , it fell completely off my radar, JOSE reminded me today, I'll address.
