### 2025-04-23 #229 @labarba
Yes, the PDF of the paper isn't compiling—that's the first thing to fix.

Could you remind us where in the Carpentries Lab are the reviews of this content posted? We agreed to fast-track already reviewed Carpentries lessons, but it would be good to leave here a link to the reviews.
