+++ md5sum.txt
@@ -3,7 +3,7 @@
 "LICENSE.md" "b24ebbb41b14ca25cf6b8216dda83e5f" "site/built/LICENSE.md" "2023-08-31"
 "config.yaml" "192dcf2d94675dfb593bc0fc2e8b2426" "site/built/config.yaml" "2024-04-24"
 "index.md" "d1e47c29006d4fd14db51a9785916fde" "site/built/index.md" "2023-08-31"
-"paper.md" "437862826f96a03774d19249189ac16c" "site/built/paper.md" "2023-09-13"
+"paper.md" "dc3fe34c287ac38ccc30c6676c312ead" "site/built/paper.md" "2025-04-23"
 "episodes/01-introduction.md" "20db3572b403389eccd97c3425717354" "site/built/01-introduction.md" "2023-08-31"
 "episodes/02-data_management.md" "7f3fd13aafe24b8b7d6b46cfa1be76a2" "site/built/02-data_management.md" "2024-03-05"
 "episodes/03-software.md" "7dc30bc0c113045c390ba3c8e4fcbf94" "site/built/03-software.md" "2023-08-31"
