### 2025-04-23 #229 @ewallace
The review process at carpentries lab is here: https://github.com/carpentries-lab/reviews/issues/24

Also a blogpost here: https://carpentries.org/blog/2024/03/good-enough-practices-carpentries-lab/

I just deleted the `date` field in commit https://github.com/carpentries-lab/good-enough-practices/commit/9bad2b1cba8230e3bdfcea520c7e63bc39329dda

@kyleniemeyer, could you let us know if that fixes the compilation?
