### 2025-04-23 2790e14
site deploy
Auto-generated via `{sandpaper}`
Source  : 8f9b57092e8b03dac14b466dd7bc0abef9ad4575
Branch  : md-outputs
Author  : GitHub Actions <actions@github.com>
Time    : 2025-04-23 18:25:33 +0000
Message : markdown source builds

Auto-generated via `{sandpaper}`
Source  : 9bad2b1cba8230e3bdfcea520c7e63bc39329dda
Branch  : main
Author  : Edward Wallace <ewallace@users.noreply.github.com>
Time    : 2025-04-23 18:24:37 +0000
Message : delete date field in paper.md

This is a minor update in a documentation markdown paper.md so I'm bypassing rules and committing changes. Doesn't change the less
[…]
