### 2025-04-23 #229 @kyleniemeyer
@ewallace the authors with multiple affilations just need quotation marks around the numbers (like `affiliation: "1, 2"`)
