### 2025-04-23 d0fecb0
quote affiliations in paper.md
This is a minor update in a documentation markdown paper.md so I'm bypassing rules and committing changes. Doesn't change the lesson content.
