### 2025-04-23 #229 @ewallace
@kyleniemeyer thanks, addressed in https://github.com/carpentries-lab/good-enough-practices/commit/d0fecb01f6d4147bd5f04ab69237436382f0dfa1
