### 2025-04-23 08929f3
markdown source builds
Auto-generated via `{sandpaper}`
Source  : d0fecb01f6d4147bd5f04ab69237436382f0dfa1
Branch  : main
Author  : Edward Wallace <ewallace@users.noreply.github.com>
Time    : 2025-04-23 18:40:53 +0000
Message : quote affiliations in paper.md

This is a minor update in a documentation markdown paper.md so I'm bypassing rules and committing changes. Doesn't change the lesson content.
