### 2025-04-23 #279 @editorialbot
[REVIEW]: Good Enough Practices in Scientific Computing: A Learning Module for Researchers
**Submitting author:** @ewallace (<a href="http://orcid.org/0000-0001-8025-6361">Edward Wallace</a>)
**Repository:** https://github.com/carpentries-lab/good-enough-practices
**Branch with paper.md** (empty if default branch):
**Version:** v1.1
