### 2025-04-23 #279 @kyleniemeyer
For the record, this has already been reviewed and approved via Carpentries Lab:

https://github.com/carpentries-lab/reviews/issues/24

https://carpentries.org/blog/2024/03/good-enough-practices-carpentries-lab/
