### 2025-04-23 #279 @labarba
Found a Zenodo deposit — is this the archive you want to associate to the JOSE publication? https://doi.org/10.5281/zenodo.10783026

Let us know if you will make a new version of the archive. Either way, we request that the title and author list match the JOSE paper.
