### 2025-04-25 6fdbcf8
Name of README.md updated to match JOSE paper
Name to Good Enough Practices in Scientific Computing: A Learning Module for Researchers
