### 2025-04-25 aa83184
Update funding in paper.md
