### 2025-04-25 5ded07a
lesson website URL fix in paper.md
