### 2025-04-25 #279 @ewallace
Thanks both! @kyleniemeyer I've made those fixes in a review thread, along with updating the funding details. https://github.com/carpentries-lab/good-enough-practices/pull/153

Were there any editorial comments or updates on the text, beyond fixing formatting issues?

@labarba we will create a new tag v1.1 with the same title. However I'm not sure how the authorship was listed for the Zenodo entry so what file I would even need to fix to address that? It looks like all the github contributors are automatically listed even if it was e.g. technical fix from Carpentries staff rather than intellectual contribution to the lesson material. Advice appreciated.
