### 2025-04-25 #279 @labarba
You can manually edit the metadata of the Zenodo archive to change the title and authors. However, if you have set up the GitHub-Zenodo integration that automatically does the archive with a new release, I believe that takes every committer and makes them an author in Zenodo. You may need to turn off the auto archive thing.
