### 2025-04-25 #279 @ewallace
I understand. Is this discrepancy acceptable, so we can accept that the zenodo archive lists everyone who has made any commit, but the JOSE paper will list the paper authors? That seems to me to make sense and to be considerably easier to achieve.
