### 2025-04-25 #279 @tobyhodges
If you add a `.zenodo.json` to the repository, [it will override the automatic extraction of authors from the list of contributors on GitHub](https://stackoverflow.com/a/57473613).
