### 2025-04-25 #279 @labarba
Thanks for the Pro Tip, @tobyhodges!

@ewallace We request that authors edit the metadata of the Zenodo deposit so title and author list match the JOSE paper. It's just cleaner that way as readers see these as part of the "same scholarly object." Could you do that?
