### 2025-04-28 #279 @tobyhodges
@ewallace I have opened [a PR to add what I believe is a valid `.zenodo.json` file](https://github.com/carpentries-lab/good-enough-practices/pull/154) to resolve the inconsistent author lists
