### 2025-04-28 #279 @labarba
Do check out the items in the author pot-review checklist [above](https://github.com/openjournals/jose-reviews/issues/279#issuecomment-2825472950).
