### 2025-04-28 #279 @ewallace
Thanks @labarba I'll do that today.

Asking again: are there any editorial comments on the text content, beyond the checklist? If it's a no that's fine, I just would like to know if there's changes to be made.
