### 2025-04-28 #279 @kyleniemeyer
@ewallace the only edits for the paper from me were shared above:
>@ewallace some paper edits:
>
> - looks like the two journal articles cited are missing DOIs, can you add those?
> - In the Summary paragraph, can you make the link at the end a proper URL? I think in Markdown that would be with < >
