### 2025-04-29 2cc4735
site deploy
Auto-generated via `{sandpaper}`
Source  : 08929f30e5f587a08d6f408482c8662ba9d51358
Branch  : md-outputs
Author  : GitHub Actions <actions@github.com>
Time    : 2025-04-23 18:41:31 +0000
Message : markdown source builds

Auto-generated via `{sandpaper}`
Source  : d0fecb01f6d4147bd5f04ab69237436382f0dfa1
Branch  : main
Author  : Edward Wallace <ewallace@users.noreply.github.com>
Time    : 2025-04-23 18:40:53 +0000
Message : quote affiliations in paper.md

This is a minor update in a documentation markdown paper.md so I'm bypassing rules and committing changes. Doesn't change the les
[…]
