+++ pkgdown.yml
@@ -1,5 +1,5 @@
 pandoc: 3.1.11
-pkgdown: 2.1.1
+pkgdown: 2.1.2
 pkgdown_sha: ~
 articles: {}
-last_built: 2025-04-23T18:41Z
+last_built: 2025-04-29T01:55Z
