### 2025-04-29 de93c9d
markdown source builds
Auto-generated via `{sandpaper}`
Source  : 057f3944571a75928fb22ae308b584c7280c7e42
Branch  : main
Author  : Edward Wallace <ewallace@users.noreply.github.com>
Time    : 2025-04-29 15:13:48 +0000
Message : Merge pull request #153 from carpentries-lab/JOSE-review-2025

Updates for JOSE review process
