### 2025-04-29 bf098a7
site deploy
Auto-generated via `{sandpaper}`
Source  : de93c9d4ab65a2ed0511cd3e085ed0ba0972c68f
Branch  : md-outputs
Author  : GitHub Actions <actions@github.com>
Time    : 2025-04-29 15:14:37 +0000
Message : markdown source builds

Auto-generated via `{sandpaper}`
Source  : 057f3944571a75928fb22ae308b584c7280c7e42
Branch  : main
Author  : Edward Wallace <ewallace@users.noreply.github.com>
Time    : 2025-04-29 15:13:48 +0000
Message : Merge pull request #153 from carpentries-lab/JOSE-review-2025

Updates for JOSE review process
