### 2025-04-29 #279 @ewallace
Thank you @kyleniemeyer and @labarba.

I've created tag v1.1 https://github.com/carpentries-lab/good-enough-practices/releases/tag/v1.1.
Update: this is now on zenodo https://doi.org/10.5281/zenodo.15304808

I believe this completes the author checklist?
