### 2025-04-29 #279 @kyleniemeyer
@ewallace the `jordan2017analysis` reference is still missing a DOI—but when I dug in, it doesn't appear that the other details are correct. That journal does not seem to have published that article... should the citation be pointing to this report archived on Zenodo instead? https://doi.org/10.5281/zenodo.1039944
