### 2025-05-02 bd5128d
Carpentries survey ref to paper.bib
Changing one reference in paper.bib, bypassing rules to commit to main.
