+++ paper.bib
@@ -23,10 +23,11 @@
 @article{jordan2017analysis,
   title={Analysis of the Carpentries' Long-Term Feedback Survey},
   author={Jordan, Kari L and Marwick, Ben and Weaver, Belinda and Zimmerman, Naupaka and Williams, Jason and Teal, Tracy and Becker, Erin and Duckles, Jonah and Duckles, Beth and Wickes, Elizabeth},
-  journal={Life Sciences},
-  volume={166},
-  pages={36--2},
-  year={2017}
+  year         = 2017,
+  publisher    = {Zenodo},
+  doi          = {10.5281/zenodo.1039944},
+  url          = {https://doi.org/10.5281/zenodo.1039944},
+
 }
 
 @misc{becker2021carpentries,
