### 2025-05-02 #279 @ewallace
@kyleniemeyer thank you, it looks like the previous reference was a Google scholar hallucination. I've fixed it and added the Zenodo reference in commit https://github.com/carpentries-lab/good-enough-practices/commit/bd5128d8815d637045f0f149507fbfef491f6908.

Does that work for you?
