### 2025-05-03 #279 @labarba
Would you like to caps-protect with `{}` the names "Carpentries" and "Software Carpentry" in the references?
