+++ paper.md
@@ -1,0 +1,148 @@
+---
+title: 'Good Enough Practices in Scientific Computing: A Learning Module for Researchers'
+tags:
+  - data
+  - computing
+  - carpentry
+  - novice
+authors:
+  - name: Tomasz Zieliński
+    email: tomasz.zielinski@ed.ac.uk
+    orcid: 0000-0002-0194-5706
+    affiliation: 1
+  - name: Andres Romanowski
+    email: andrew.romanowski@ed.ac.uk
+    orcid: 0000-0003-0737-2408
+    affiliation: "1,2"
+  - name: Emma Wilson
+    email: emma.wilson@ed.ac.uk
+    orcid: 0000-0002-8100-7508
+    affiliation: 3
+  - name: Felicity Anderson
+    email: felicity.anderson@ed.ac.uk
+    orcid: 0000-0001-8778-6779
+    affiliation: "1,4"
+  - name:  Elvina Gountouna
+    email: e.gountouna@ed.ac.uk
+    orcid: 0000-0001-7870-2780
+    affiliation: "5, 6"
+  - name: Matthias Mimault
+    email: matthias.mimault@hutton.ac.uk
+    orcid: 0000-0003-1516-7332
+    affiliation: 7
+  - name: Alison Meynert
+    email: alison.meynert@ed.ac.uk
+    orcid: 0000-0001-5839-1751
+    affiliation: 5
+  - name: Edward W.J. Wallace
+    email: edward.wallace@ed.ac.uk
+    orcid: 0000-0001-8025-6361
+    affiliation: 1
+affiliations:
+ - name: School of Biological Sciences, The University of Edinburgh
+   index: 1
+ - name: Plant-Environment Signaling Group, Dept. of Biology, Faculty of Science, Utrecht University, Utrecht, 3584 CH, The Netherlands
+   index: 2
+ - name: Centre for Clinical Brain Sciences, The University of Edinburgh
+   index: 3
+ - name: EPCC, The University of Edinburgh
+   index: 4
+ - name: MRC Human Genetics Unit, MRC Institute of Genetics and Cancer, The University of Edinburgh
+   index: 5
+ - name: Usher Institute, The University of Edinburgh
+   index: 6
+ - name: The James Hutton Institute, Dundee
+   index: 7
+bibliography: paper.bib
+
+
+---
+
+# Summary
+
+We present a half-day learning module targeted at a broad audience of researchers who want to learn how to be more efficient and effective in their data analysis and computing, whatever their career stage.
+The module teaches "good enough practices" that are near-universally useful for researchers who use computers in their work.
+These practices encompass data management, software and programming, collaborating with colleagues, organizing projects, keeping track of changes, and writing manuscripts.
+Good enough practices rely on a shared set of principles that span these areas: planning, modular organization, names, and documentation.
+The lesson is in The Carpentries format and the materials are open-source and hosted on GitHub by The Carpentries Lab.
+The lesson is visible at <https://carpentries-lab.github.io/good-enough-practices/>.
+
+
+
+# Statement of Need
+
+<!-- explain how the submitted artifacts contribute to computationally enabled teaching and learning, and describing how they might be adopted by others. -->
+
+This lesson addresses good practices for data analysis, computing, and collaboration, that are broadly useful for researchers.
+The underlying principles of planning, modular organization, good names, and documentation, are important for both novices and experts who take computational approaches.
+This learning module helps novices to learn these good practices in their own right, empowering them to learn other computational skills.
+For example, novice programmers can struggle with poor planning, poorly organized files and projects, unclear names for variables, documenting and describing their code.
+Teaching good practices separately means that programming lessons can focus on programming concepts and syntax.
+Making good practices explicit also helps more advanced researchers to collaborate with others, and to explicitly teach good practice to novices.
+
[…]
