+++ md5sum.txt
@@ -1,0 +1,21 @@
+"file" "checksum" "built" "date"
+"CODE_OF_CONDUCT.md" "c93c83c630db2fe2462240bf72552548" "site/built/CODE_OF_CONDUCT.md" "2023-08-31"
+"LICENSE.md" "b24ebbb41b14ca25cf6b8216dda83e5f" "site/built/LICENSE.md" "2023-08-31"
+"config.yaml" "192dcf2d94675dfb593bc0fc2e8b2426" "site/built/config.yaml" "2024-04-24"
+"index.md" "d1e47c29006d4fd14db51a9785916fde" "site/built/index.md" "2023-08-31"
+"paper.md" "3330fdb0675e8c3cbb2439f0dc0505e0" "site/built/paper.md" "2025-04-29"
+"episodes/01-introduction.md" "20db3572b403389eccd97c3425717354" "site/built/01-introduction.md" "2023-08-31"
+"episodes/02-data_management.md" "7f3fd13aafe24b8b7d6b46cfa1be76a2" "site/built/02-data_management.md" "2024-03-05"
+"episodes/03-software.md" "7dc30bc0c113045c390ba3c8e4fcbf94" "site/built/03-software.md" "2023-08-31"
+"episodes/04-collaboration.md" "2bd0f509102efb43ad311055a381e475" "site/built/04-collaboration.md" "2023-09-23"
+"episodes/05-project_organization.md" "c2a58b5bbc3e45493582c1527b3f82e2" "site/built/05-project_organization.md" "2023-09-23"
+"episodes/06-track_changes.md" "b02e8243730f0192d4418b6843fc72e0" "site/built/06-track_changes.md" "2023-09-07"
+"episodes/07-manuscripts.md" "cd54d77e18d0e00bdd0ad41343e461ad" "site/built/07-manuscripts.md" "2023-08-31"
+"episodes/08-what_next.md" "d7f89e1953957185712810e48661387a" "site/built/08-what_next.md" "2023-08-31"
+"instructors/instructor-notes.md" "07295dd92ccf82cdea8b86442cfecfb8" "site/built/instructor-notes.md" "2023-08-31"
+"instructors/etherpad_template.md" "af9be0a87fcb6b6ea41bd229e6157cf9" "site/built/etherpad_template.md" "2023-09-07"
+"instructors/teaching_guide_for_etherpad.md" "4368b3c5c4be6bd82246ba24dcfbbfd5" "site/built/teaching_guide_for_etherpad.md" "2023-09-07"
+"instructors/resources.md" "034c8ddb0dc4caa5226a34ec6a6284a3" "site/built/resources.md" "2023-08-31"
+"instructors/what-we-left-out.md" "86af63461877aa02ab3f10c2b8170ccb" "site/built/what-we-left-out.md" "2023-08-31"
+"learners/setup.md" "bc7c787f5667e01309870dd8ff627aaa" "site/built/setup.md" "2023-08-31"
+"profiles/learner-profiles.md" "a207bdd02b3f3dc7b82d2d4d17f03a6e" "site/built/learner-profiles.md" "2023-09-07"
