+++ config.yaml
@@ -1,0 +1,86 @@
+#------------------------------------------------------------
+# Values for this lesson.
+#------------------------------------------------------------
+
+# Which carpentry is this (swc, dc, lc, or cp)?
+# swc: Software Carpentry
+# dc: Data Carpentry
+# lc: Library Carpentry
+# cp: Carpentries (to use for instructor training for instance)
+# incubator: The Carpentries Incubator
+# lab: The Carpentries Lab
+carpentry: 'lab'
+
+# Overall title for pages.
+title: 'Good Enough Practices in Scientific Computing'
+
+# Date the lesson was created (YYYY-MM-DD, this is empty by default)
+created: '2020-08-24'
+
+# Comma-separated list of keywords for the lesson
+keywords: 'software, data, lesson, The Carpentries'
+
+# Life cycle stage of the lesson
+# possible values: pre-alpha, alpha, beta, stable
+life_cycle: 'stable'
+
+# License of the lesson materials (recommended CC-BY 4.0)
+license: 'CC-BY 4.0'
+
+# Link to the source repository for this lesson
+source: 'https://github.com/carpentries-lab/good-enough-practices'
+
+# Link to doi for the lesson on zenodo
+doi: '10.5281/zenodo.10783026'
+
+
+# Default branch of your lesson
+branch: 'main'
+
+# Who to contact if there are any issues
+contact: 'Edward.Wallace@ed.ac.uk'
+
+# Navigation ------------------------------------------------
+#
+# Use the following menu items to specify the order of
+# individual pages in each dropdown section. Leave blank to
+# include all pages in the folder.
+#
+
+# Order of episodes in your lesson
+episodes:
+- 01-introduction.md
+- 02-data_management.md
+- 03-software.md
+- 04-collaboration.md
+- 05-project_organization.md
+- 06-track_changes.md
+- 07-manuscripts.md
+- 08-what_next.md
+
+# Information for Learners
+learners:
+- setup.md
+
+# Information for Instructors
+instructors:
+- instructor-notes.md
+- etherpad_template.md
+- teaching_guide_for_etherpad.md
+- resources.md
+- what-we-left-out.md
+
+# Learner Profiles
+profiles:
+- learner-profiles.md
+
+
+# Customisation ---------------------------------------------
+#
+# This space below is where custom yaml items (e.g. pinning
+# sandpaper and varnish versions) should live
+
+
+url: 'https://carpentries-lab.github.io/good-enough-practices'
+analytics: carpentries
+lang: en

[…]
