### 2025-05-05 #279 @ewallace
Thanks @labarba. In the process of cross-checking this case-protecion I found some other references to Carpentries lessons that need updatng

Working on this in pull request https://github.com/carpentries-lab/good-enough-practices/pull/155, and I've asked @emma-wilson to review

Two questions for you.

1. When The Carpentries citation file requests citing with editors names only (e.g. https://github.com/swcarpentry/git-novice/blob/main/CITATION), is it ok to fill in the "editor" field not the "author" field in the bibtex?
2. We referenced the Carpentries curriculum development handbook, now deprecated (https://cdh.carpentries.org). Should we update to the Collaborative Lesson Development Training (https://carpentries.github.io/lesson-development-training/), which has a doi, as a better reference, even though the material was in a different form when we referred to it?

Sorry this is taking so many steps!
Edward
