### 2025-05-05 #279 @labarba
1. You need to try it with the "editor" field and see if it renders properly. If it does not, you could add the editors in the"author" field with the string "(eds.)" strategically placed (i.e., a hack).
2. You can reference the updated source, and perhaps in the paper text mention the now-deprecated handbook, maybe in a parenthetical.
