### 2025-05-12 #279 @ewallace
Thanks for the reminder @labarba !

I've updated the references including "editor" field at https://github.com/carpentries-lab/good-enough-practices/commit/42dd9943fb19d1cc431557ea319bbd47713c95c7

Could we check if this compiles properly?
