### 2025-05-12 #279 @ewallace
Now update the lesson dev handbook reference.
