### 2025-05-12 cfb2e0a
site deploy
Auto-generated via `{sandpaper}`
Source  : e3d4374db05c51493735cc7617f6ae572e27c41d
Branch  : md-outputs
Author  : GitHub Actions <actions@github.com>
Time    : 2025-05-12 09:06:14 +0000
Message : markdown source builds

Auto-generated via `{sandpaper}`
Source  : f5c7ff7543d02c812815fe3370f46b291928ea8e
Branch  : main
Author  : Edward Wallace <ewallace@users.noreply.github.com>
Time    : 2025-05-12 08:56:32 +0000
Message : Merge pull request #156 from carpentries-lab/lesson-dev

Lesson Dev update in paper.bib and paper.md

Just an update to paper /citations (again) so not waiting f
[…]
