### 2025-05-12 618c3d0
markdown source builds
Auto-generated via `{sandpaper}`
Source  : 265084aff9cb83af1aecf9b7d6df3f31ea9123f8
Branch  : main
Author  : Edward Wallace <ewallace@users.noreply.github.com>
Time    : 2025-05-12 09:32:53 +0000
Message : More fix refs in paper.md
