### 2025-05-12 a615170
More formatting fixes in paper.bib
