+++ paper.bib
@@ -55,7 +55,7 @@
                   Via, Allegra and
                   Wilson, Greg and
                   Word, Karen},
-  title        = {The Carpentries Collaborative Lesson Development
+  title        = {{T}he {C}arpentries Collaborative Lesson Development
                    Training
                   },
   year         = 2024,
