### 2025-05-12 c68a4cc
site deploy
Auto-generated via `{sandpaper}`
Source  : 618c3d012eb4b3f2f3effb93d4e32e78593edfa3
Branch  : md-outputs
Author  : GitHub Actions <actions@github.com>
Time    : 2025-05-12 09:33:46 +0000
Message : markdown source builds

Auto-generated via `{sandpaper}`
Source  : 265084aff9cb83af1aecf9b7d6df3f31ea9123f8
Branch  : main
Author  : Edward Wallace <ewallace@users.noreply.github.com>
Time    : 2025-05-12 09:32:53 +0000
Message : More fix refs in paper.md
